+++ src/Kiota.Builder/Refiners/CSharpReservedClassNamesProvider.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Kiota.Builder.Refiners;
+
+public class CSharpReservedClassNamesProvider : IReservedNamesProvider
+{
+    private readonly Lazy<HashSet<string>> _reservedNames = new(static () => new(StringComparer.OrdinalIgnoreCase)
+    {
+        "BaseRequestBuilder",
+        "BaseCliRequestBuilder"
+    });
+    public HashSet<string> ReservedNames => _reservedNames.Value;
+}