+++ tests/Kiota.Builder.Tests/Writers/CSharp/CodeMethodWriterTests.cs
@@ -582,6 +582,108 @@ public class CodeMethodWriterTests : IDisposable
         AssertExtensions.CurlyBracesAreClosed(result);
     }
     [Fact]
+    public void WritesModelFactoryBodyAndDisambiguateAmbiguousImportedTypes()
+    {
+        // Arrange : Adding a model with conflicting discriminator
+        var modelsNamespace = root.AddNamespace("models");
+        var modelWithConflictingDiscriminatorTypes = modelsNamespace.AddClass(
+            new CodeClass
+            {
+                Name = "ConflictingModelBaseClass"
+            }).First();
+        modelsNamespace.AddClass(modelWithConflictingDiscriminatorTypes);
+        // Adding a model in another namespace discriminator
+        var levelOneNameSpace = modelsNamespace.AddNamespace("namespaceLevelOne");
+        var conflictingModelInOtherNamespace = levelOneNameSpace.AddClass(
+            new CodeClass
+            {
+                Name = "ConflictingModel",
+                StartBlock = new ClassDeclaration()
+                {
+                    Inherits = new CodeType()
+                    {
+                        Name = modelWithConflictingDiscriminatorTypes.Name,
+                        TypeDefinition = modelWithConflictingDiscriminatorTypes
+                    }
+                }
+
+            }).First();
+        levelOneNameSpace.AddClass(conflictingModelInOtherNamespace);
+        // Adding an enum in another namespace that conflicts
+        var levelTwoNameSpace = modelsNamespace.AddNamespace("namespaceLevelTwo");
+        var conflictingEnumInOtherNamespace = levelTwoNameSpace.AddEnum(
+            new CodeEnum()
+            {
+                Name = "ConflictingModel"
+            }).First();
+        modelsNamespace.AddEnum(conflictingEnumInOtherNamespace);// add enum to root models with same name
+
+        // setup the usings
+        modelWithConflictingDiscriminatorTypes.AddUsing(
+            new CodeUsing()
+            {
+                Name = levelOneNameSpace.Name,
+                Declaration = new CodeType()
+                {
+                    Name = conflictingModelInOtherNamespace.Name,
+                    TypeDefinition = conflictingModelInOtherNamespace
+                },
+            });
+        modelWithConflictingDiscriminatorTypes.AddUsing(
+            new CodeUsing()
+            {
+                Name = levelTwoNameSpace.Name,
+                Declaration = new CodeType()
+                {
+                    Name = conflictingEnumInOtherNamespace.Name,
+                    TypeDefinition = conflictingEnumInOtherNamespace
+                },
+            });
+        modelWithConflictingDiscriminatorTypes.DiscriminatorInformation = new DiscriminatorInformation()
+        {
+            Name = "type",
+            DiscriminatorPropertyName = "@odata.type"
+        };
+        modelWithConflictingDiscriminatorTypes.DiscriminatorInformation.AddDiscriminatorMapping(levelOneNameSpace.Name + ".ConflictingModel", new CodeType()
+        {
+            Name = conflictingModelInOtherNamespace.Name,
+            TypeDefinition = conflictingModelInOtherNamespace
+        });
+
+        var factoryMethod = modelWithConflictingDiscriminatorTypes.AddMethod(new CodeMethod
+        {
+            Name = "factory",
+            Kind = CodeMethodKind.Factory,
+            ReturnType = new CodeType
+            {
+                Name = "parentModel",
+                TypeDefinition = modelWithConflictingDiscriminatorTypes,
+            },
+            IsStatic = true,
+        }).First();
+
+        factoryMethod.AddParameter(new CodeParameter
+        {
+            Name = "parseNode",
+            Kind = CodeParameterKind.ParseNode,
+            Type = new CodeType
+            {
+                Name = "ParseNode",
+                IsExternal = true,
+            },
+            Optional = false,
+        });
+
+        writer.Write(factoryMethod);
+        var result = tw.ToString();
+
+        Assert.Contains("var mappingValue = parseNode.GetChildNode(\"@odata.type\")?.GetStringValue()", result);
+        Assert.Contains("return mappingValue switch {", result);
+        Assert.Contains("\"namespaceLevelOne.ConflictingModel\" => new namespaceLevelOne.ConflictingModel(),", result); //Assert the disambiguation happens due to the enum imported
+        Assert.Contains("_ => new ConflictingModelBaseClass()", result);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
     public void WritesModelFactoryBodyForInheritedModels()
     {
         var parentModel = root.AddClass(new CodeClass
@@ -798,9 +900,9 @@ public class CodeMethodWriterTests : IDisposable
         Assert.Contains("var requestConfig = new RequestConfig()", result);
         Assert.Contains("config.Invoke(requestConfig)", result);
         Assert.Contains("requestInfo.Headers.Add(\"Accept\", \"application/json\")", result);
-        Assert.Contains("requestInfo.AddHeaders(requestConfig.H)", result);
-        Assert.Contains("requestInfo.AddQueryParameters(requestConfig.Q)", result);
-        Assert.Contains("requestInfo.AddRequestOptions(requestConfig.O)", result);
+        Assert.Contains("requestInfo.AddHeaders(requestConfig.Headers)", result);
+        Assert.Contains("requestInfo.AddQueryParameters(requestConfig.QueryParameters)", result);
+        Assert.Contains("requestInfo.AddRequestOptions(requestConfig.Options)", result);
         Assert.Contains("SetContentFromScalar", result);
         Assert.Contains("return requestInfo;", result);
         AssertExtensions.CurlyBracesAreClosed(result, 1);
@@ -824,9 +926,9 @@ public class CodeMethodWriterTests : IDisposable
         Assert.Contains("var requestConfig = new RequestConfig()", result);
         Assert.Contains("config.Invoke(requestConfig)", result);
         Assert.Contains("requestInfo.Headers.Add(\"Accept\", \"application/json\")", result);
-        Assert.Contains("requestInfo.AddHeaders(requestConfig.H)", result);
-        Assert.Contains("requestInfo.AddQueryParameters(requestConfig.Q)", result);
-        Assert.Contains("requestInfo.AddRequestOptions(requestConfig.O)", result);
+        Assert.Contains("requestInfo.AddHeaders(requestConfig.Headers)", result);
+        Assert.Contains("requestInfo.AddQueryParameters(requestConfig.QueryParameters)", result);
+        Assert.Contains("requestInfo.AddRequestOptions(requestConfig.Options)", result);
         Assert.Contains("SetContentFromScalar", result);
         Assert.Contains("return requestInfo;", result);
         Assert.Contains("async Task<double?>", result);//verify we only have one nullable marker
@@ -851,9 +953,9 @@ public class CodeMethodWriterTests : IDisposable
         Assert.Contains("var requestConfig = new RequestConfig()", result);
         Assert.Contains("config.Invoke(requestConfig)", result);
         Assert.Contains("requestInfo.Headers.Add(\"Accept\", \"application/json\")", result);
-        Assert.Contains("requestInfo.AddHeaders(requestConfig.H)", result);
-        Assert.Contains("requestInfo.AddQueryParameters(requestConfig.Q)", result);
-        Assert.Contains("requestInfo.AddRequestOptions(requestConfig.O)", result);
+        Assert.Contains("requestInfo.AddHeaders(requestConfig.Headers)", result);
+        Assert.Contains("requestInfo.AddQueryParameters(requestConfig.QueryParameters)", result);
+        Assert.Contains("requestInfo.AddRequestOptions(requestConfig.Options)", result);
         Assert.Contains("SetContentFromParsable", result);
         Assert.Contains("return requestInfo;", result);
         AssertExtensions.CurlyBracesAreClosed(result, 1);
@@ -1229,7 +1331,7 @@ public class CodeMethodWriterTests : IDisposable
         {
             Name = propName,
             DefaultValue = defaultValue,
-            Kind = CodePropertyKind.UrlTemplate,
+            Kind = CodePropertyKind.Custom,
             Type = new CodeType
             {
                 Name = "string"
@@ -1427,7 +1529,7 @@ public class CodeMethodWriterTests : IDisposable
         method.AddParameter(new CodeParameter
         {
             Name = "ra",
-            Kind = CodeParameterKind.RequestAdapter,
+            Kind = CodeParameterKind.Custom,
             Type = new CodeType
             {
                 Name = "RequestAdapter",
@@ -1461,6 +1563,37 @@ public class CodeMethodWriterTests : IDisposable
     }
 
     [Fact]
+    public void ConstructorCallsCliBaseClass()
+    {
+        method.ReturnType = new CodeType
+        {
+            Name = "void",
+            IsExternal = true
+        };
+        method.Kind = CodeMethodKind.Constructor;
+        parentClass.Kind = CodeClassKind.RequestBuilder;
+        parentClass.StartBlock.Inherits = new CodeType
+        {
+            Name = "BaseCliRequestBuilder",
+            IsExternal = true
+        };
+        parentClass.AddProperty(new CodeProperty
+        {
+            Type = new CodeType
+            {
+                Name = "string",
+                IsExternal = true
+            },
+            Kind = CodePropertyKind.UrlTemplate,
+            DefaultValue = "\"test\"",
+            Name = "urlTpl"
+        });
+        writer.Write(method);
+        var result = tw.ToString();
+        Assert.Contains(": base(\"test\")", result);
+    }
+
+    [Fact]
     public void WritesMethodWithEmptyStringAsDefaultValueIfNotNullableAndOptional()
     {
         method.ReturnType = new CodeType
@@ -1472,7 +1605,7 @@ public class CodeMethodWriterTests : IDisposable
         method.AddParameter(new CodeParameter
         {
             Name = "ra",
-            Kind = CodeParameterKind.RequestAdapter,
+            Kind = CodeParameterKind.Custom,
             Type = new CodeType
             {
                 Name = "RequestAdapter",