+++ CHANGELOG.md
@@ -9,13 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added Visual Studio Code preview extension. [#2333](https://github.com/microsoft/kiota/issues/2333)
 - Added support for searching in forks for API descriptions in GitHub. [#2429](https://github.com/microsoft/kiota/issues/2429)
+- Added the ability to filter on operations. [#2431](https://github.com/microsoft/kiota/issues/2431)
 
 ### Changed
 
+- Fixed a bug where unescaped query parameters are not correctly matched to the original name due to python convention of snake casing parameter names. 
+- Fixed a bug where date types annotations and guid's were not correctly translated in Python
+- Fixed the extension of downloaded files when using the default path. [#2316](https://github.com/microsoft/kiota/issues/2316)
 - Fixed a bug where lookup of reference ids failed for AllOf more than one level up.
 - Fixed a bug where a CLI client would not set the content types for requests. (Shell)
 - Fixed linting errors by re-ordering methods and properties in Python.
+- Fixed a bug where requests with bodies would fail in Ruby. [#2473](https://github.com/microsoft/kiota/issues/2473)
 - Changed python import mechanism to facilitate code completion. [#2380](https://github.com/microsoft/kiota/issues/2380)
 - Fixed a bug where discriminator methods were missing possible types in Python [#2381](https://github.com/microsoft/kiota/issues/2381)
 - Fixed a bug where boolean or number enums would be mapped to enums instead of primitive types. [#2367](https://github.com/microsoft/kiota/issues/2367)
@@ -27,7 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug where java writer would emit incorrect type names in case of compound types
 - Fixed a bug where go refiner would emit incorrect code when inlining error parents
 - Fixed a bug in PHP where the base URL path parameter key didn't match the URI template.
-
+- Changed invalid schema behavior, invalid properties are now ignored with a warning instead of failed. [#2089](https://github.com/microsoft/kiota/issues/2089)
+- Fixed a bug where java refiner would not normalize inherited class names and interface types.
+- Fixed a bug where search based commands would not match exact matches when additional results are available.
+- Fixed a bug where imported classes and enums would not be disambiguated when they have the same name in dotnet.
 
 ## [1.0.1] - 2023-03-11